The unreachable branch goes; the derived guard already made the claim
`wowRendersDefault.test.tsx` looped `SURFACE_KEYS` and branched on a
hand-maintained `WOW_SKINNED` set built from that same list. The set had grown
to hold all twenty keys, so the `else` could not run — and a NEW surface key
defaulted to *unclaimed*, i.e. the file would have asserted WOW falls back on
it, the opposite of what it exists to guard.

`surfaceDispatch.test.ts` makes the same claim derived: `REQUIRED` is every
surface all five reference factions skin, and it is currently all twenty keys —
the same set — so `describe('WOW is bespoke on every core surface')` covers
every row the loop did. Verified red before deleting: de-registering `vote`
from `wow.ts` (a surface with no row in that file's `BESPOKE` table, so only
the derived describe can catch it) fails `wow skins vote`.

The theme half survives in `wowIsThemed.test.ts`, named for what it tests: the
hue tokens resolve to declared names, `isKnownFaction('wow')` is true, and WOW
has a name and description of its own rather than falling through to `names.na`
and labelling a populated faction "Unaffiliated".

The two manifest docblocks that pointed at the old file were stale on a second
count too — both still said WOW is "themed-and-partly-skinned, not fully
dressed" and that the rest fall back. They now point at the derived guard.

Closes #2534
